Updated after Data interface change.

+++ chart/chistoxrange.cpp
@@ -7,7 +7,7 @@
 
 namespace chart {
 
-void CHistoXRange::init(const IDataColumn *c, size_t bc)
+void CHistoXRange::init(const IDataColumn *c, int bc)
 {
     Q_ASSERT(c != 0);
     m_nBinCount = bc;
@@ -17,7 +17,7 @@ void CHistoXRange::init(const IDataColumn *c, size_t bc)
     m_fBinSize = av;
 }
 
-size_t CHistoXRange::getSize() const
+int CHistoXRange::getSize() const
 {
     return m_nBinCount;
 }
@@ -29,7 +29,7 @@ QPair<double, double> CHistoXRange::getRange() const
     return QPair<double, double>(x, y);
 }
 
-void CHistoXRange::getData(size_t row, double &val) const
+void CHistoXRange::getData(int row, double &val) const
 {
     val = m_fBinAnchor + (row + 0.5) * m_fBinSize;
 }
