+++ chart/chistoxrange.h
@@ -19,7 +19,7 @@ public:
      * @param[in] c pointer to IDatacoulmn for wich to create range
      * @param[in] bc bin count for histogram chart
      */
-    void init(const data::IDataColumn* c, size_t bc = 10);
+    void init(const data::IDataColumn* c, int bc = 10);
 
     /// @name Reimplemented interface
     /// @{
@@ -27,14 +27,14 @@ public:
     virtual QPair<double, double> getRange() const;
 
 public:
-    virtual size_t getSize() const;
+    virtual int getSize() const;
 
 public:
-    virtual void getData(size_t row, double &val) const;
+    virtual void getData(int row, double &val) const;
     /// @}
 
 private:
-    size_t m_nBinCount;
+    int m_nBinCount;
     double m_fBinSize;
     double m_fBinAnchor;
 };
