+++ chart/chistoyrange.h
@@ -20,7 +20,7 @@ public:
      * @param[in] c column for which to create range
      * @param[in] bc bins count
      */
-    void init(const data::IDataColumn* c, size_t bc);
+    void init(const data::IDataColumn* c, int bc);
 
     /// @name Reimplemented interfaces
     /// @{
@@ -28,15 +28,15 @@ public:
     virtual QPair<double, double> getRange() const;
 
 public:
-    virtual size_t getSize() const;
+    virtual int getSize() const;
 
 public:
-    virtual void getData(size_t row, double &val) const;
+    virtual void getData(int row, double &val) const;
     /// @}
 
 private:
     int m_nBinCount;
-    QVector<size_t> m_vecBins;
+    QVector<int> m_vecBins;
 };
 
 } // namespace chart
